scripts/requirements.txt
# Dependencies for spec generation scripts (e.g. generate_outline.py)
# Install: pip install -r scripts/requirements.txt
PyYAML>=6.0
